requirements.txt
Jinja2>=3.0.2
selenium>=3.141.0
msedge-selenium-tools>=3.141.3